打榜/ev_detection/requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

tensorboard>=2.4.1

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

onnx>=1.9.0  # ONNX export
onnx-simplifier>=0.3.6  # ONNX simplifier
scikit-learn==0.19.2  # CoreML quantization

thop  # FLOPs computation